Moved the core "robot control" elements to a separate "robot" folder.
